# swagger.yaml
openapi: 3.0.0
info:
  title: DentApp API
  description: API для стоматологического онлайн-сервиса. Позволяет управлять пользователями, записью на приём, расписанием врачей, историей посещений и предварительной диагностикой.
  version: 1.0.0
  contact:
    name: Команда DentApp
    email: support@dentapp.ru
servers:
  - url: http://localhost:3000/api
    description: Локальный сервер разработки
  - url: https://api.dentapp.ru/v1
    description: Продакшн сервер
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Введите JWT токен, полученный при входе
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Иван Иванов"
        email:
          type: string
          format: email
          example: "ivan@example.com"
        phone:
          type: string
          example: "+7 (123) 456-78-90"
        role:
          type: string
          enum: [patient, doctor, admin]
          example: "patient"
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
    Doctor:
      type: object
      properties:
        id:
          type: integer
          example: 2
        name:
          type: string
          example: "Петров Алексей Владимирович"
        specialty:
          type: string
          example: "Терапевт, хирург"
        experience_years:
          type: integer
          example: 12
        rating:
          type: number
          format: float
          example: 4.8
        bio:
          type: string
          example: "Врач высшей категории, стаж 12 лет"
      required:
        - id
        - name
        - specialty
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
          example: 101
        patient_id:
          type: integer
          example: 1
        doctor_id:
          type: integer
          example: 2
        service_id:
          type: integer
          example: 1
        appointment_date:
          type: string
          format: date
          example: "2023-08-15"
        appointment_time:
          type: string
          format: time
          example: "10:00"
        status:
          type: string
          enum: [scheduled, completed, cancelled]
          example: "scheduled"
        notes:
          type: string
          example: "Просьба не задерживаться"
        created_at:
          type: string
          format: date-time
      required:
        - id
        - patient_id
        - doctor_id
        - service_id
        - appointment_date
        - appointment_time
        - status
...
    LoginRequest:
      type: object
      required:
        - credential
        - password
      properties:
        credential:
          type: string
          description: Email или телефон пользователя
          example: "ivan@example.com"
        password:
          type: string
          format: password
          example: "secret123"
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        user:
          $ref: '#/components/schemas/User'
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - phone
        - password
      properties:
        name:
          type: string
          example: "Иван Иванов"
        email:
          type: string
          format: email
          example: "ivan@example.com"
        phone:
          type: string
          example: "+71234567890"
        password:
          type: string
          format: password
          minLength: 6
          example: "secret123"
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          example: "Иван Петров"
        phone:
          type: string
          example: "+70987654321"
        email:
          type: string
          format: email
          example: "new@example.com"
    AppointmentCreate:
      type: object
      required:
        - doctor_id
        - service_id
        - date
        - time
      properties:
        doctor_id:
          type: integer
          example: 2
        service_id:
          type: integer
          example: 1
        date:
          type: string
          format: date
          example: "2023-08-15"
        time:
          type: string
          format: time
          example: "10:00"
        notes:
          type: string
          example: "Просьба не задерживаться"
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Описание ошибки"
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: Вход в систему
      description: Аутентификация пользователя по email/телефону и паролю. Возвращает JWT токен.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      tags:
        - Auth
      summary: Регистрация нового пользователя
      description: Создаёт учётную запись пациента. Роль по умолчанию — patient.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Ошибка валидации (например, email уже занят)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/profile:
    get:
      tags:
        - Auth
      summary: Получить профиль текущего пользователя
      description: Возвращает информацию об авторизованном пользователе.
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Не авторизован
    put:
      tags:
        - Auth
      summary: Обновить профиль текущего пользователя
      description: Изменяет имя, телефон или email. Email должен быть уникальным.
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Профиль обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Ошибка валидации
        '401':
          description: Не авторизован

  # ========== USERS (admin only) ==========
...
  /doctors:
    get:
      tags:
        - Doctors
      summary: Список врачей
      description: Возвращает список всех врачей с возможностью фильтрации по специальности.
      operationId: getDoctors
      parameters:
        - name: specialty
          in: query
          schema:
            type: string
          description: Фильтр по специальности (например, "терапевт")
      responses:
        '200':
          description: Список врачей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Doctor'
...
  /appointments:
    get:
      tags:
        - Appointments
      summary: Список записей текущего пользователя
      description: Для пациента возвращает его записи, для врача — все записи к нему, для администратора — все записи.
      operationId: getAppointments
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, completed, cancelled]
          description: Фильтр по статусу
        - name: from_date
          in: query
          schema:
            type: string
            format: date
          description: Начало периода
        - name: to_date
          in: query
          schema:
            type: string
            format: date
          description: Конец периода
      responses:
        '200':
          description: Список записей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
        '401':
          description: Не авторизован

    post:
      tags:
        - Appointments
      summary: Создать запись на приём
      description: Создаёт новую запись. Проверяет доступность слота.
      operationId: createAppointment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCreate'
      responses:
        '201':
          description: Запись создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          description: Неверные данные
        '401':
          description: Не авторизован
        '409':
          description: Слот уже занят
...